+++ src/application/VcsPanoramaFooter.vue
@@ -8,7 +8,9 @@
   const time = ref();
   const hasLinks = computed(
     () =>
-      !!app.uiConfig.config?.dataProtection || !!app.uiConfig.config?.imprint,
+      !!app.uiConfig.config?.dataProtection ||
+      !!app.uiConfig.config?.imprint ||
+      app.uiConfig.config?.footerInformation?.length > 0,
   );
 
   const showPanoramaFooter = computed(